pyproject.toml
[tool.poetry]
name = "cat-charity-fund"
version = "0.1.0"
description = "Приложение для Благотворительного фонда поддержки котиков QRKot. Фонд собирает пожертвования на различные целевые проекты: на медицинское обслуживание нуждающихся хвостатых, на обустройство кошачьей колонии в подвале, на корм оставшимся без попечения кошкам — на любые цели, связанные с поддержкой кошачьей популяции."
authors = ["Roman <flexomj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.78.0"
uvicorn = { version = "0.17.6", extras = ["standard"] }
pydantic = "1.10.0"
aiosqlite = "0.17.0"
sqlalchemy = "1.4.36"
alembic = "1.7.7"
fastapi-users-db-sqlalchemy = "4.0.3"
requests = "2.27.1"
passlib = "^1.7.4"
pyparsing = "3.0.9"


[tool.poetry.group.dev.dependencies]
pytest = "7.1.3"
pytest-asyncio = "0.23.4"
pytest-freezegun = "0.4.2"
pytest-lazy-fixture = "0.6.3"
pytest-pythonpath = "0.7.3"
black = "^24.4.2"
ruff = "^0.5.3"
flake8 = "4.0.1"
mixer = "^7.2.2"
toml = "^0.10.2"
anyio = "3.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79


